refactor: make circle abstract series
