+++ src/component/bubbleSeries.ts
@@ -1,17 +1,9 @@
-import Component from './component';
 import { CircleModel } from '@t/components/series';
-import { BaseOptions, BubbleSeriesType, CoordinateDataType } from '@t/options';
-import { ClipRectAreaModel } from '@t/components/series';
+import { BaseOptions, BubbleSeriesType } from '@t/options';
 import { ChartState, Scale, SeriesTheme } from '@t/store/store';
-import { TooltipData } from '@t/components/tooltip';
-import {
-  getCoordinateDataIndex,
-  getCoordinateLabel,
-  getCoordinateValue
-} from '@src/helpers/coordinate';
+import { getCoordinateLabel, getCoordinateValue } from '@src/helpers/coordinate';
 import { getRGBA } from '@src/helpers/color';
-
-type DrawModels = ClipRectAreaModel | CircleModel;
+import CircleSeries from '@src/component/circleSeries';
 
 interface RenderOptions {
   theme: SeriesTheme;
@@ -19,27 +11,14 @@ interface RenderOptions {
 
 const MINIMUM_DETECTING_AREA_RADIUS = 1;
 
-// @TODO: scatter series와 하나가 될 수 있을까?
-export default class BubbleSeries extends Component {
-  models!: DrawModels[];
-
-  responders!: CircleModel[];
-
-  activatedResponders: this['responders'] = [];
-
+export default class BubbleSeries extends CircleSeries {
   maxRadius = -1;
 
   maxValue = -1;
 
   initialize() {
-    this.type = 'series';
+    super.initialize();
     this.name = 'bubbleSeries';
-  }
-
-  update(delta: number) {
-    if (this.models[0].type === 'clipRectArea') {
-      this.models[0].width = this.rect.width * delta;
-    }
   }
 
   setMaxValue(bubbleData: BubbleSeriesType[]) {
@@ -55,8 +34,10 @@ export default class BubbleSeries extends Component {
     }
 
     const { xAxis, yAxis } = axes;
-
     const bubbleData = series.bubble.data;
+    const renderOptions: RenderOptions = {
+      theme: theme.series
+    };
 
     this.rect = layout.plot;
 
@@ -64,14 +45,9 @@ export default class BubbleSeries extends Component {
     const yAxisTickSize = this.rect.height / yAxis!.tickCount;
 
     this.maxRadius = Math.min(xAxisTickSize, yAxisTickSize);
-
-    const renderOptions: RenderOptions = {
-      theme: theme.series
-    };
-
     this.setMaxValue(bubbleData);
-    const seriesModel = this.renderBubblePointsModel(bubbleData, renderOptions, scale);
 
+    const seriesModel = this.renderBubblePointsModel(bubbleData, renderOptions, scale);
     const tooltipModel = this.makeTooltipModel(bubbleData, categories, renderOptions);
 
     this.models = [this.renderClipRectAreaModel(), ...seriesModel];
@@ -84,39 +60,6 @@ export default class BubbleSeries extends Component {
       style: ['default', 'hover', { lineWidth: 2 }],
       data: tooltipModel[index]
     }));
-  }
-
-  makeTooltipModel(
-    bubbleData: BubbleSeriesType[],
-    categories: string[],
-    renderOptions: RenderOptions
-  ) {
-    const { theme } = renderOptions;
-
-    return bubbleData.flatMap(({ data, name }, index) => {
-      const tooltipData: TooltipData[] = [];
-
-      data.forEach((datum: CoordinateDataType, dataIdx) => {
-        tooltipData.push({
-          label: name,
-          color: theme.colors[index],
-          value: getCoordinateValue(datum),
-          category: categories[getCoordinateDataIndex(datum, categories, dataIdx)]
-        });
-      });
-
-      return tooltipData;
-    });
-  }
-
-  renderClipRectAreaModel(): ClipRectAreaModel {
-    return {
-      type: 'clipRectArea',
-      x: 0,
-      y: 0,
-      width: 0,
-      height: this.rect.height
-    };
   }
 
   renderBubblePointsModel(
@@ -159,24 +102,5 @@ export default class BubbleSeries extends Component {
 
       return circleModels;
     });
-  }
-
-  onMousemove({ responders }) {
-    this.activatedResponders.forEach(responder => {
-      const index = this.models.findIndex(model => model === responder);
-      this.models.splice(index, 1);
-    });
-
-    responders
-      .sort((a: CircleModel, b: CircleModel) => b.radius - a.radius)
-      .forEach(responder => {
-        this.models.push(responder);
-      });
-
-    this.activatedResponders = responders;
-
-    this.eventBus.emit('seriesPointHovered', this.activatedResponders);
-
-    this.eventBus.emit('needDraw');
   }
 }
